Move the state into the base class.
